+++ tests/banking/test_routes.py
@@ -2,7 +2,7 @@
 import json
 
 import pytest
-from authanor.testing.helpers import transaction_lifetime
+from fuisce.testing import transaction_lifetime
 
 from test_helpers import TestRoutes